chore: refactor project
