+++ server/server.ts
@@ -3,9 +3,9 @@ dotEnvConfig()
 // set up log level
 import './setup'
 
-import { initBroker } from 'infra/messaging/rabbit-mq-messaging'
+import { initBroker } from 'messaging/broker'
 
-import registerEventHandlers from './services/event-handlers'
+import registerEventHandlers from './services/eventHandlers'
 
 async function main() {
   const broker = await initBroker()
