+++ server/services/knex.ts
@@ -1,12 +1,10 @@
 import { Knex, knex } from 'knex'
 import log from 'loglevel'
 
-const connection = process.env.DATABASE_URL
-
 const knexConfig: Knex.Config = {
   client: 'pg',
   debug: process.env.NODE_LOG_LEVEL === 'debug',
-  connection,
+  connection: process.env.DATABASE_URL,
   pool: { min: 0, max: 10 },
 }
